reduced hdf5 complexity
