+++ nac/schedule/hdf5_interface.py
@@ -1,12 +1,7 @@
 """Module to store data in the HDF5."""
 
-__all__ = ['store_cp2k_basis']
-from os.path import join
-
 import h5py
 import numpy as np
-
-from qmflows.parsers.cp2KParser import readCp2KBasis
 
 
 class StoreasHDF5:
@@ -36,36 +31,3 @@ class StoreasHDF5:
         """
         dset = self.save_data(path_property, data)
         dset.attrs[name_attr] = attr
-
-    def saveBasis(self, parser_fun: callable, path_basis: str):
-        """Store the basis set.
-
-        :param parser_fun: Function to parse the file containing the
-                          information about the primitive contracted Gauss
-                          functions.
-        :param path_basis: Absolute path to the file containing the basis
-                          sets information.
-        :returns: None
-        """
-        keys, vals = parser_fun(path_basis)
-        pathsExpo = [join("cp2k/basis", xs.atom, xs.basis, "exponents")
-                     for xs in keys]
-        pathsCoeff = [join("cp2k/basis", xs.atom, xs.basis,
-                           "coefficients") for xs in keys]
-
-        for ps, es in zip(pathsExpo, [xs.exponents for xs in vals]):
-            self.save_data(ps, es)
-
-        fss = [xs.basisFormat for xs in keys]
-        css = [xs.coefficients for xs in vals]
-
-        # save basis set coefficients and their correspoding format
-        for path, fs, css in zip(pathsCoeff, fss, css):
-            self.save_data_attrs("basisFormat", str(fs), path, css)
-
-
-def store_cp2k_basis(file_h5, path_basis):
-    """Read the CP2K basis set into an HDF5 file."""
-    storeCp2k = StoreasHDF5(file_h5)
-
-    return storeCp2k.saveBasis(readCp2KBasis, path_basis)
